Initial Commit
This is the initial commit. The current version should be able to run in
a web browser so you can play the game without downloading all the
files. However, due to new browser restrictions, most browsers won't
allow you to play it for security reasons. Because of that, I am
currently working on adapting it back for desktop use.